fix: resolve 12 bugs — registry deadlock, config deep-merge, MPS selection, VAD race, and more
- Fix BackendRegistry deadlock: RLock replaces Lock, _backends check replaces list_backends() in register_builtin_backends
- Fix TOCTOU race in is_available(): capture backend_class inside lock before releasing
- Fix SecretStr comparison in get_backend(): new _configs_match() compares secret values field-by-field
- Fix config merger: TOML sub-models now deep-merge with defaults instead of replacing entire dict
- Fix auto-selection preferring faster-whisper on MPS: _has_gpu() renamed to _has_cuda(), MPS excluded since CTranslate2 falls back to CPU
- Fix VAD reset_state() race: acquires _inference_lock before resetting model state
- Fix temp file leak in extract_audio(): document caller responsibility, add logger warning
- Fix subprocess.wait() timeout: add 30s timeout on final process.wait() after stderr streaming
- Fix falsy model_size fallback: use explicit None check instead of truthy check
- Fix OpenAIConfig.timeout type: changed from int to float for OpenAI client compatibility
- Add regression tests for all fixes

+++ src/audiocore/config/openai_config.py
@@ -57,10 +57,10 @@ class OpenAIConfig(BaseModel):
         default=None,
         description="Custom base URL for API requests (optional, for proxies)",
     )
-    timeout: int = Field(
-        default=300,
-        ge=1,
-        le=3600,
+    timeout: float = Field(
+        default=300.0,
+        ge=1.0,
+        le=3600.0,
         description="Request timeout in seconds (default 300 for large files)",
     )
     max_retries: int = Field(
